+++ Cmdlets/Public/Remove-TeamViewerPSProxy.ps1
@@ -4,9 +4,9 @@ function Remove-TeamViewerPSProxy {
 
     $global:TeamViewerProxyUriRemoved = $true
     $global:TeamViewerProxyUriRemoved | Out-Null  # https://github.com/PowerShell/PSScriptAnalyzer/issues/1472
-    if($PSCmdlet.ShouldProcess($TeamViewerProxyUriRemoved,"Remove proxy for WebAPI")){
-    $global:TeamViewerProxyUriSet = $null
-    $global:TeamViewerProxyUriSet | Out-Null  # https://github.com/PowerShell/PSScriptAnalyzer/issues/1472
-    [Environment]::SetEnvironmentVariable('TeamViewerProxyUri','', 'User')
+    if ($PSCmdlet.ShouldProcess($TeamViewerProxyUriRemoved, 'Remove proxy for web API')) {
+        $global:TeamViewerProxyUriSet = $null
+        $global:TeamViewerProxyUriSet | Out-Null  # https://github.com/PowerShell/PSScriptAnalyzer/issues/1472
+        [Environment]::SetEnvironmentVariable('TeamViewerProxyUri', '', 'User')
     }
 }